+++ Example/RMPhotoLibrary/ViewController.swift
@@ -1,0 +1,24 @@
+//
+//  ViewController.swift
+//  RMPhotoLibrary
+//
+//  Created by reiya.matsuki.1990.09.23@gmail.com on 06/06/2017.
+//  Copyright (c) 2017 reiya.matsuki.1990.09.23@gmail.com. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+
+}
+
