--- PhotoLibrary/ViewController.swift
@@ -1,25 +1,0 @@
-//
-//  ViewController.swift
-//  PhotoLibrary
-//
-//  Created by Matsuki, Reiya (松木 玲也) on 2017/06/06.
-//  Copyright © 2017年 reiya. All rights reserved.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view, typically from a nib.
-    }
-
-    override func didReceiveMemoryWarning() {
-        super.didReceiveMemoryWarning()
-        // Dispose of any resources that can be recreated.
-    }
-
-
-}
-
