+++ README.md
@@ -1,0 +1,29 @@
+# RMPhotoLibrary
+
+[![CI Status](http://img.shields.io/travis/reiya.matsuki.1990.09.23@gmail.com/RMPhotoLibrary.svg?style=flat)](https://travis-ci.org/reiya.matsuki.1990.09.23@gmail.com/RMPhotoLibrary)
+[![Version](https://img.shields.io/cocoapods/v/RMPhotoLibrary.svg?style=flat)](http://cocoapods.org/pods/RMPhotoLibrary)
+[![License](https://img.shields.io/cocoapods/l/RMPhotoLibrary.svg?style=flat)](http://cocoapods.org/pods/RMPhotoLibrary)
+[![Platform](https://img.shields.io/cocoapods/p/RMPhotoLibrary.svg?style=flat)](http://cocoapods.org/pods/RMPhotoLibrary)
+
+## Example
+
+To run the example project, clone the repo, and run `pod install` from the Example directory first.
+
+## Requirements
+
+## Installation
+
+RMPhotoLibrary is available through [CocoaPods](http://cocoapods.org). To install
+it, simply add the following line to your Podfile:
+
+```ruby
+pod "RMPhotoLibrary"
+```
+
+## Author
+
+reiya.matsuki.1990.09.23@gmail.com, t-rmatsuki@valuecommerce.co.jp
+
+## License
+
+RMPhotoLibrary is available under the MIT license. See the LICENSE file for more info.
